refactor: 사용하지 않는 컴포넌트 삭제
