--- components/LoginButton/index.tsx
@@ -1,3 +1,0 @@
-import LoginButton from './LoginButton';
-
-export default LoginButton;
